src/requirements.txt
alembic==1.5.4
click==7.1.2
Flask==1.1.2
Flask-Admin==1.5.7
Flask-Login==0.5.0
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.4.1
six==1.15.0
SQLAlchemy==1.3.23
Werkzeug==1.0.1
WTForms==2.3.3
